[PLAT-483] refactor app routes into separate
